backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic[email]
pydantic-settings==2.3.4
aiomysql==0.2.0
python-multipart==0.0.9
greenlet==3.2.4
cryptography==45.0.6
passlib==1.7.4
python-jose==3.5.0
bcrypt==3.2.2